refactor(protocol,openomni): move hashline to openomni plan module
