+++ packages/openomni/src/plan/hashline.ts
