+++ packages/protocol/src/index.ts
@@ -15,4 +15,3 @@ export * from "./event-log/index.js";
 export * from "./agent/index.js";
 export * from "./artifact/index.js";
 export * from "./gate/index.js";
-export * from "./hashline/index.js";
